feat: file-manager add module utils.mjs
